Make get.sh generic for cross repo sharing
Currently the get.sh scripts across a number of repos are roughly based off the faas-cli filewith bespoke changes made depending upon the needs of the respective projects.

This change make the faas-cli script generic enough so that the common content can be used across all the repos without any changes.  The only changes required to tailor the file's operation to the respective repo are to be made in the Repo specific content section.

Also adds a check for a git bash signature to enable downloading of the .exe variant when running on git bash on Windows.

Signed-off-by: Richard Gee <richard@technologee.co.uk>